removes a big bottleneck
This artifact did an insert+commit for every result found. Changed it to gather every result in a list, and then do a db.executemany() + db.commit() insert with all the results, per file.

This speeds up the artifact ~100x, and the total time of a full scan goes down 50-60%.