minor optimizations on walstrings artifact
Instead of writing every line as it matches things, output lines are kept in a list and then a single file.write() call is done, if results were found.

Also replaces re.finditer() for re.findall(), which fits the use case of walstrings nicely and brings a very small speed improvement (about 2-3% faster, mainly from avoiding m.group() calls).

All combined make for a 10-12% speed improvement in the artifact. Explored cython as an alternative to the regex, but that's an entirely different discussion on its own.