Update Health.py
This change parses data entered by the user of the device through the Fitness Application (Personalize Fitness and Health) for both height and weight. Once parsed both height and weight are displayed with the timestamp the data was entered followed by height in meters, centimeters, and feet and inches and weight in kilograms, stones, and pounds.