Update vipps.py
Updated Vipps Transactions due to the BLOB/JSON in database is different if it's a message or a payment. 

A new quick update to the numbers as it required a . to divide the amount. Vipps does not accept cents (øre), but also there is a limit of how much one can send per transaction before paying a fee. Something many avoid :)