Add files via upload
This provides an in-depth review of all Apple devices and their OS version timeframes parsed from the Health Application. Adds parsing support for Device and OS Timeframe, Device Isolation (both iOS 15, 16, 17, and 18), and Current Model and OS (iOS 18 only) for Apple devices, pulling data from the healthdb_secure and healthdb databases. 
Additional information is available here: https://metadataperspective.com/2025/01/30/beyond-the-logs-using-the-health-app-to-uncover-device-model-and-os-history/