Create callHistoryGroupCall
Group Calls were not previously supported in parsing, this parses group calls from the CallHistory.storedata and interactionC.db. Given the correct dataset for testing, the CallHistory.storedata portion of this could be added to the callHistory.py artifact. More information on the research pertaining to this work is available here: https://metadataperspective.com/2025/02/05/hello-who-is-on-the-line/